refactor(auth): actions folder for alt before api calls for auth & clean up folders/files
