+++ web/app/(landing)/error/page.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 const ErrorPage = () => {
   return <p>Sorry, something went wrong</p>;
 };
